requirements.txt
uvicorn[standard]
fastapi
sqlalchemy
databases[aiosqlite]
python-dotenv
rich
asgi-correlation-id
python-json-logger
python-jose
python-multipart
passlib[bcrypt]
httpx